3078 수정, 1966 풀이

+++ Zojae031/CodingTest/3078.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 int main() {
 	int n, k;
-	int cnt = 0;
+	long long cnt = 0;
 	int deleteIndex = 0;
 	cin >> n>> k;
 	getchar();
